+++ i18n/example_test.go
@@ -8,6 +8,7 @@ package i18n
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
+	"bytes"
 	"fmt"
 	"strings"
 )
@@ -143,6 +144,30 @@ func ExampleText_With() {
 	// Output:
 	// Unknown user johndoe
 	// Unknown ID 183
+}
+
+func ExampleText_Write() {
+	en := &Bundle{
+		GREETING: "Hello!",
+		ERRORS: &Errors{
+			UNKNOWN_USER: "Unknown user {{.Username}}",
+			UNKNOWN_ID:   "Unknown ID {{.ID}}",
+		},
+	}
+
+	data := Data{
+		"Username": "johndoe",
+		"ID":       183,
+	}
+
+	buf := &bytes.Buffer{}
+
+	en.ERRORS.UNKNOWN_USER.Write(buf, data)
+
+	fmt.Println(buf.String())
+
+	// Output:
+	// Unknown user johndoe
 }
 
 func ExampleText_Add() {